remove sftp config
