--- .vscode/sftp.json
@@ -1,50 +1,0 @@
-{
-    "name": "deepray",
-    "protocol": "sftp",
-    "port": 22,
-    "username": "root",
-    "uploadOnSave": true,
-    "watcher": {
-        "files": "**/*",
-        "autoUpload": true,
-        "autoDelete": true
-    },
-    "syncOption": {
-        "delete": true
-    },
-    "profiles": {
-        "A100": {
-            "host": "10.0.74.1",
-            "username": "fuhailin",
-            "password": "Ab123456",
-            "remotePath": "/data/fuhailin/workspaces/deepray"
-        },
-        "TITAN": {
-            "host": "10.0.74.1",
-            "username": "root",
-            "password": "Arsenaldevk8sTITAN",
-            "remotePath": "/data/fuhailin/workspaces/deepray"
-        },
-        "A30": {
-            "host": "10.0.74.11",
-            "username": "root",
-            "password": "actest",
-            "remotePath": "/data/fuhailin/workspaces/deepray"
-        }
-    },
-    "defaultProfile": "A30",
-    "ignore": [
-        ".vscode",
-        ".DS_Store",
-        "project_details.txt",
-        ".eslintrc.json",
-        "node_modules",
-        ".idea",
-        "__pycache__",
-        ".pyc",
-        "bazel-*",
-        ".tf_configure.bazelrc",
-        "wheels*",
-        ".git"
-    ]
-}
